This repository includes files for the ALY6030 Final Project on pharmacy claims.
This project focuses on normalizing pharmacy claims data and setting up a star schema for analysis. The provided CSV files include the normalized data in 3NF, separated into fact and dimension tables. The SQL script details the creation of the database schema, including primary and foreign keys, as well as sample queries for reporting and analysis. The final report provides a comprehensive overview of the project, including answers to questions on fact types, primary and foreign keys, and considerations for foreign key constraints.